backend/backend/requirements.txt
asgiref==3.11.1
brotli==1.2.0
dj-database-url==3.1.2
Django>=5.0,<6.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
packaging==26.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
